Updated MoveVehicle and MoveSingleVehicle to make them function properly. Before the change they would not put the correct format in the new string index. Had to also slightly edit the FindVehicle method in order to do this.
Changed PairSingleMcs method to not be a loop. As there can only be one new single Mc per loop of the main menu, the method does not need to be a loop.

Changed the output format of the ShowParkingSpaces method to make it all fit on one console screen.

Removed the AlreadyRegistered method as it was more or less exactly the same as FindVehicle. Changed ParkVehicle to call on FindVehicle instead.

Made sure the outputs were all correct.